tests: Move mapped storage coverage
Mapped storage now lives under core. The storage primitive tests should
follow the package that owns byte storage and record vectors, while batch
tests should cover only the matcher workspace wrapper.

The old test module mixed primitive storage behavior with MatcherWorkspace
ownership checks. That made package placement harder to read after
mapped_storage moved out of utils.

This commit addresses that by placing mapped-storage primitive coverage in
tests/core/test_mapped_storage.py and leaving the batch matcher workspace
check in tests/batch/test_match_storage.py.

Architecture coverage for the package boundary will follow in a separate
tests commit.
